requirements.txt
discord.py==0.16.11
urllib3